Struct bomberman::game_data::GameData

pub struct GameData {
    pub bombs: Vec<Bomb>,
    pub enemies: Vec<Enemy>,
    pub detours: Vec<Detour>,
    pub maze: Vec<Vec<String>>,
    pub wall_interceps: bool,
    pub rock_interceps: bool,
}

Fields
bombs: Vec<Bomb>enemies: Vec<Enemy>detours: Vec<Detour>maze: Vec<Vec<String>>wall_interceps: boolrock_interceps: bool
Implementations

impl GameData

Crea una nueva instancia de GameData con los datos especificados.

Argumentos

bombs: Vector de bombas en el juego.

enemies: Vector de enemigos en el juego.

detours: Vector de detours en el juego.

maze: Laberinto del juego representado como una matriz de cadenas.

wall_interceps: Indica si las intercepciones de muros están habilitadas.

rock_interceps: Indica si las intercepciones de rocas están habilitadas.

pub fn new( bombs: Vec<Bomb>, enemies: Vec<Enemy>, detours: Vec<Detour>, maze: Vec<Vec<String>>, wall_interceps: bool, rock_interceps: bool ) -> Self

pub fn find_bomb( &mut self, coordinate_x: usize, coordinate_y: usize ) -> Option<&mut Bomb>

Encuentra una bomba en las coordenadas especificadas y devuelve una referencia mutable a ella.

Argumentos

coordinate_x: Coordenada X de la bomba.

coordinate_y: Coordenada Y de la bomba.

pub fn remove_bomb(&mut self, coordinate_x: usize, coordinate_y: usize)

Elimina una bomba en las coordenadas especificadas.

Argumentos

coordinate_x: Coordenada X de la bomba a eliminar.

coordinate_y: Coordenada Y de la bomba a eliminar.

pub fn check_paths<'a>( new_x: usize, game_data: &'a mut GameData, y: usize, typee: TypeBomb, bomb: &'a Bomb, iterations_pending: usize )

Esta función verifica los caminos posibles y aplica efectos en la dirección hacia abajo en el juego.

Argumentos

new_x: La nueva coordenada X a la que se moverá la función.

game_data: Una referencia mutable a los datos del juego.

y: La coordenada Y actual en la que se encuentra la función.

typee: El tipo de bomba que se está utilizando.

bomb: Una referencia a la bomba que se está utilizando.

iterations_pending: El número de iteraciones pendientes en la dirección hacia abajo.

pub fn move_down<'a>( game_data: &'a mut GameData, x: usize, y: usize, reach: usize, typee: TypeBomb, bomb: &'a Bomb ) -> &'a mut GameData

Esta función mueve al personaje hacia abajo en el laberinto y aplica efectos de bombas u obstáculos en su camino.

Argumentos

game_data: Una referencia mutable a los datos del juego.

x: La coordenada X actual del personaje.

y: La coordenada Y actual del personaje.

reach: La distancia máxima que el personaje puede moverse hacia abajo.

typee: El tipo de bomba que se está utilizando.

bomb: Una referencia a la bomba que se está utilizando.

Devolución

Devuelve una referencia mutable a los datos del juego actualizados después de mover al personaje.

pub fn apply_bomb_effect( &mut self, coordinate_x: usize, coordinate_y: usize, reach: usize, tipo_bomb: TypeBomb, copy_bomb: &Bomb )

Esta función aplica el efecto de una bomba en las cuatro direcciones (arriba, abajo, izquierda y derecha) desde la posición especificada.

Argumentos

coordinate_x: La coordenada X de la posición desde la cual se aplicará el efecto de la bomba.

coordinate_y: La coordenada Y de la posición desde la cual se aplicará el efecto de la bomba.

reach: La distancia máxima que el efecto de la bomba puede alcanzar en cada dirección.

tipo_bomb: El tipo de bomba que se está utilizando.

copy_bomb: Una referencia a la bomba que se está utilizando.

pub fn validate_maze( &self, coordinate_x: usize, coordinate_y: usize ) -> Result<(), Box<dyn Error>>

Esta función valida el estado actual del laberinto en las coordenadas especificadas.

Argumentos

coordinate_x: La coordenada X a validar.

coordinate_y: La coordenada Y a validar.

Retorna

Ok(()): Si se encuentra una bomba válida en las coordenadas especificadas y si hay enemigos con vidas válidas cerca.

Err: Si no se cumplen las condiciones de validación, se devuelve un error con detalles sobre objetos inválidos.

pub fn handle_detour( game_data: &mut GameData, object: &String, new_x: usize, y: usize, typee: TypeBomb, iterations_pending: usize, bomb: &Bomb )

Esta función maneja una situación en la que se encuentra un objeto de desvío en el laberinto.

Argumentos

game_data: Una referencia mutable a los datos del juego.

object: Una cadena que representa el tipo de objeto de desvío (“DU”, “DD”, “DR”, o “DL”).

new_x: La nueva coordenada X después de seguir el objeto de desvío.

y: La coordenada Y actual.

typee: El tipo de bomba utilizada para seguir el objeto de desvío.

iterations_pending: El número de iteraciones pendientes para seguir el objeto de desvío.

bomb: Una referencia a la bomba actual.

pub fn handle_enemy( game_data: &mut GameData, object: &String, new_x: usize, y: usize, typee: TypeBomb, iterations_pending: usize, bomb: &Bomb )

Esta función maneja una situación en la que se encuentra un enemigo en el laberinto.

Argumentos

game_data: Una referencia mutable a los datos del juego.

object: Una cadena que representa el tipo de enemigo.

new_x: La nueva coordenada X después de encontrarse con el enemigo.

y: La coordenada Y actual.

typee: El tipo de bomba utilizada para enfrentar al enemigo.

iterations_pending: El número de iteraciones pendientes después de enfrentarse al enemigo.

bomb: Una referencia a la bomba utilizada para enfrentarse al enemigo.

pub fn handle_rock(game_data: &mut GameData)

Esta función establece la señal de que un objeto tipo “rock” ha interceptado el camino.

Argumentos

game_data: Una referencia mutable a los datos del juego.

pub fn handle_wall(game_data: &mut GameData)

Esta función establece la señal de que un objeto tipo “wall” ha interceptado el camino.

Argumentos

game_data: Una referencia mutable a los datos del juego.

pub fn handle_bomb( game_data: &mut GameData, object: &String, new_x: usize, y: usize )

Esta función maneja una situación en la que se encuentra una bomba en el laberinto.

Argumentos

game_data: Una referencia mutable a los datos del juego.

object: Una cadena que representa el tipo de objeto en la posición actual (e.g., “B” para bomba).

new_x: La nueva coordenada X después de encontrarse con la bomba.

y: La coordenada Y actual.

pub fn move_up<'a>( game_data: &'a mut GameData, x: usize, y: usize, reach: usize, typee: TypeBomb, bomb: &'a Bomb ) -> &'a mut GameData

Mueve al jugador hacia arriba en el laberinto hasta alcanzar una distancia máxima especificada o hasta encontrar un obstáculo.

Esta función se encarga de mover al jugador hacia arriba en el laberinto, una casilla a la vez, hasta alcanzar una distancia máxima especificada o hasta encontrar un obstáculo como una pared o una roca. Para cada paso, verifica si la nueva posición está dentro de los límites del laberinto y realiza las comprobaciones necesarias llamando a check_paths. Si se encuentra con una pared o una roca, se detiene. Al final, restablece las señales de intercepción de pared y roca y devuelve una referencia mutable a los datos del juego actualizados.

Argumentos

game_data: Una referencia mutable a los datos del juego.

x: La coordenada X actual del jugador.

y: La coordenada Y actual del jugador.

reach: La distancia máxima que puede moverse hacia arriba.

typee: El tipo de bomba que se utiliza.

bomb: Una referencia a la bomba utilizada.

Devuelve

Una referencia mutable a los datos del juego actualizados después del movimiento.

pub fn move_right<'a>( game_data: &'a mut GameData, x: usize, y: usize, reach: usize, typee: TypeBomb, bomb: &'a Bomb ) -> &'a mut GameData

Mueve al jugador hacia la derecha en el laberinto hasta alcanzar una distancia máxima especificada o hasta encontrar un obstáculo.

Esta función se encarga de mover al jugador hacia la derecha en el laberinto, una casilla a la vez, hasta alcanzar una distancia máxima especificada o hasta encontrar un obstáculo como una pared o una roca. Para cada paso, verifica si la nueva posición está dentro de los límites del laberinto y realiza las comprobaciones necesarias llamando a check_paths. Si se encuentra con una pared o una roca, se detiene. Al final, restablece las señales de intercepción de pared y roca y devuelve una referencia mutable a los datos del juego actualizados.

Argumentos

game_data: Una referencia mutable a los datos del juego.

x: La coordenada X actual del jugador.

y: La coordenada Y actual del jugador.

reach: La distancia máxima que puede moverse hacia la derecha.

typee: El tipo de bomba que se utiliza.

bomb: Una referencia a la bomba utilizada.

Devuelve

Una referencia mutable a los datos del juego actualizados después del movimiento.

pub fn move_left<'a>( game_data: &'a mut GameData, x: usize, y: usize, reach: usize, typee: TypeBomb, bomb: &'a Bomb ) -> &'a mut GameData

Mueve al jugador hacia la izquierda en el laberinto hasta alcanzar una distancia máxima especificada o hasta encontrar un obstáculo.

Esta función se encarga de mover al jugador hacia la izquierda en el laberinto, una casilla a la vez, hasta alcanzar una distancia máxima especificada o hasta encontrar un obstáculo como una pared o una roca. Para cada paso, verifica si la nueva posición está dentro de los límites del laberinto y realiza las comprobaciones necesarias llamando a check_paths. Si se encuentra con una pared o una roca, se detiene. Al final, restablece las señales de intercepción de pared y roca y devuelve una referencia mutable a los datos del juego actualizados.

Argumentos

game_data: Una referencia mutable a los datos del juego.

x: La coordenada X actual del jugador.

y: La coordenada Y actual del jugador.

reach: La distancia máxima que puede moverse hacia la izquierda.

typee: El tipo de bomba que se utiliza.

bomb: Una referencia a la bomba utilizada.

Devuelve

Una referencia mutable a los datos del juego actualizados después del movimiento.

pub fn print_maze(maze: &Vec<Vec<String>>)

Imprime el laberinto en la consola.

Esta función toma una referencia a una matriz bidimensional de cadenas que representa el laberinto y lo imprime en la consola. Cada elemento de la matriz se imprime sin espacio adicional entre ellos, y se agrega un espacio en blanco después de cada celda para separar las columnas. Además, se agrega un salto de línea al final de cada fila para separar las filas en la salida.

Argumentos

maze: Una referencia a la matriz bidimensional que representa el laberinto.